Data Engineer Task/requirements.txt
flask
flask-sqlalchemy
Faker
pandas
pymysql